### 2025-05-17 #8212 @editorialbot
[PRE REVIEW]: eVaiutilities: Data management utilities for eVai
**Submitting author:** @IBCHgenomic (<a href="http://orcid.org/0000-0002-4157-9405">Gaurav Sablok</a>)
**Repository:** https://github.com/IBCHgenomic/eVaiutilities
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
