### 2025-05-18 #8212 @IBCHgenomic
@arfon Thank you for letting me know and i appreciate this for letting me know. If i can do anything to help the paper process faster then i would like to do the same.

I would deeply appreciate as if JOSS can help review the paper faster.

Thank you for for serving as a Editor in Chief and helping developing scientific communities.

Gaurav
