### 2025-06-10 #8212 @lskatz
Hi, thank you for considering me; however, I do not consider human genomics in my realm of expertise.
