### 2025-06-12 #8212 @matthewfeickert
@IBCHgenomic I am not the associate editor for your submission (c.f. https://github.com/openjournals/joss-reviews/issues/8212#issuecomment-2956655055). Please be patient and let the managing track EiC continue with the process of finding an editor who can take on this proposal at the moment.
