### 2025-08-15 #8212 @jromanowska
Hi, @codebiogenomics  :wave:
I will take care of the process from now on. I will try to find reviewers.
In the meantime, please add DOIs to the reference list in your papers and format the URLs as hyper-links (i.e., so they're clickable).
