### 2025-08-22 #8212 @genomicssport
@jromanowska thank you for reviewing the paper, I have added some new additions such as faster sqlx and also a variant accumulator with plots ready to be used for inference. Now it is 10 times faster than the previous version and uses end to end asynchronous code.
