### 2025-08-22 #8212 @erik-whiting
> :wave: @zhengxinchang, @sigven & @erik-whiting, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'm happy to review. I've done a few reviews for JOSS before and I think the project is in my wheelhouse.
