### 2025-08-22 #8212 @jromanowska
Thank you, @erik-whiting, for agreeing to review!
@genomicssport - great that you're constantly improving the software! Just a small clarification: I am not a reviewer, I am editor and it's actually me who can start the review process.
