### 2025-08-25 #8853 @genomicssport
@jromanowska @erik-whiting @JMencius I just finished and added code and it now also plots the figure also: https://github.com/genomicssport/eVaiutilities. Previously it was doing only the statistical analysis and now it plots all the variant analysis. So 1 command variant accumulate and it analyzes thousand of the eVai files in less than a minute and output the figure for the analysis. No need of any plotting library.

Thank you for making the paper review faster and i thank you all and JOSS for the timely review, which reflects the academic integration and thinking. I hope that JOSS is not biased and is purely scientific evaluation.
