### 2025-08-26 #8853 @genomicssport
@JMencius Thank you so much for adding that way and actually suggesting that way. I appreciate you for suggesting that. I wanted to put a comparative picture of usage eVAI over years as it has been cited and used a lot but then I went through the JOSS papers and guidelines and it seems they have kept strictly computational, so I kept that for a another release build to add to the GitHub repository.

You are absolutely right and that is why eVaiUtilities has a large scope for all those who are using eVai.

Thank you all for this discussion and I am open to review and answer any comments pertaining to the paper.
