### 2025-08-26 #8853 @genomicssport
@jromanowska thank you for the comment. eVaiutilities is the first one to address all the additional analysis, so i dont see a point where a paper can use it before it is published. So i am not quite sure about your comment: could you point to an overview or a paper where eVai functionality is listed and specifically say what your software enables that eVai can't do?.

The eVaiutilities does at large scale and at command line which makes it easier for the access to the people who want to analyze the large scale datasets and faster. eVai is web based and has no command line functions and these functionalities are not present.
