### 2025-08-26 #8853 @jromanowska
Maybe it wasn't clear. I understand that your software gives functionality that is not in eVai. However, none of us (reviewers or me) have used eVai and the link that is in your paper doesn't lead to much information - it's more an advertisement than actual list of functionality. Your software doesn't have much documentation either. Is there a tutorial or some examples to learn about what each function does? Is it really so that eVai doesn't have possibility to filter the results?
