### 2025-08-26 #8853 @genomicssport
> > 3. eVAI does the variant annnotation, however it does on a sample basis means you can analyze one sample, but not a large scale of populations. Imagine sitting in a open browser and running continuously without outputting results. so eVaiutilities resolves this. It comes at lightning speed and presents all the analysis at population scale. Furthermore, it has several options such as alt-allele analysis, variant accumulate, which eVAI lacks and user have to do custom analysis at their end. So eVaiutilities solve all those and present you a comprehensive suite of analysis utilities that makes your analysis scalable and high-throughput.
>
> [@genomicssport](https://github.com/genomicssport) Thank you for the explaination, so I kind of understand now, so `eVaiutilities` is not entirely dependent on the result output by `eVai`, it can independently handle output from standard format file (`.tsv`) as you shown in your example in https://github.com/IBCHgenomic/eVaiutilities. It is more like a professional CLI and updated version of `eVai`. Am I correct?

@JMencius you are absolutely right. In the next version, I am adding the https://www.crimsoniris.com/ version also, so that it can give you personal genomics front end also.
