### 2025-08-26 #8853 @JMencius
@genomicssport Now I know it, so maybe you should rephase the sentence in the Abstract, which is quite confusing:

> eVaiutilities proposed in this paper, is a data management software for the analysis of the eVai output files.

I will install and run the software in days and finish this review ASAP.
