### 2025-08-26 #8853 @genomicssport
@JMencius Thank you for this suggestion and i will do that now and will make it a post-eVAI analyzer for population scale analysis.
