### 2025-08-26 #8853 @jromanowska
@genomicssport - yes, there was a misunderstanding and thank you for clarification. Please, update the README to clarify that your software is not dependent on eVai output.

What I meant with more documentation is that each function should have a description and all arguments listed, as well as maybe a small example code _with output_. I don't know Rust, so I'm not sure what is the standard there, but e.g., in R, you can even have a webpage with documentation page for each function - look e.g., here: https://vatlab.github.io/vat-docs/documentation/functions/ref_sequence/
