### 2025-08-26 #8853 @genomicssport
@jromanowska Thanks you and i have already done that. I have updated the README with more documentation and how it should be run. Thank you for these comments and happy to answer any additional comment if you have any for reviewing. I also code in Python and R and React, So i am well versed with these all programming languages.

I was a proficient C++ developer and turned to RUST(Everything) and now only uses RUST, Python and React(Typescript) for everything. PyTorch and Tensorflow for machine learning and R only for some analysis.
