### 2025-08-28 #8853 @JMencius
> [@genomicssport](https://github.com/genomicssport), please add DOIs to the references in your paper and make the links clickable.

@genomicssport   please add DOI links to the references, such as
```
Pedersen, B. S., & Ridder, J. de. (2022). Echtvar: Compressed variant representation for 70 rapid annotation and filtering of SNPs and indels. Nucleic Acids Research, 51(1), e3–e3.1 <mark>https://doi.org/10.1093/nar/gkac931</mark>
```
