### 2025-08-31 #8853 @erik-whiting
Hello, I updated my review to check off everything except the references and Documentation. Am I understanding correctly that the documentation is currently being updated? If so, I will finish my review of it when updates are completed. If not, I review the existing documentation as is.

As a note, I did check off the performance-based check:
>Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

As far as I can tell, the software makes no performance claims so I don't think this part is necessary:

> Hi [@genomicssport](https://github.com/genomicssport) Please also add expected runtime or memory usage (maybe just for test samples and list the dimension of the test sample). You can take repository https://github.com/JMencius/LongBow?tab=readme-ov-file#resource-consumption as a reference.
>
> ```
> Performance: If there are any performance claims of the software, have they been confirmed?
> ```
>
> and I will check it while trying.

Great work so far! I was particularly impressed with the improvements to the article itself in the last few days.
