### 2025-09-02 #8853 @genomicssport
@erik-whiting Thank you so much for the good words. I am deeply impressed by having a fair review and scientific words. Very good. The performance gain is that i added async across all the functions, so that it becomes faster.

I am going to finish all the corrections this evening or may be tomorrow.

Brilliant and fair review and scientifically discussed by you. Hook me in if you are submitting to JOSS or developing RUST or learning RUST.

Good regards, Gaurav
