### 2025-09-05 #8853 @JMencius
@genomicssport   And also please specify the Rust version you used to develop as a reference in `README.md`. I used a relatively old version of Rust will trigger
```
cargo install eVaiutilities
Updating crates.io index
Downloaded eVaiutilities v0.2.0 error: failed to parse manifest at /home/jzlab/.cargo/registry/src/index.crates.io-6f17d22bba15001f/eVaiutilities-0.2.0/Cargo.toml Caused by: feature edition2024 is required The package requires the Cargo feature called edition2024, but that feature is not stabilized in this version of Cargo (1.79.0 (ffa9cf99a 2024-06-03)). Consider trying a newer version of Cargo (this may require the nightly release). See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#edition-2024 for more information about the status of this feature
```
