### 2025-09-08 #8853 @danielskatz
> I wonder if the `Reviewer` part can show my ORCID in the manuscript

JOSS does not currently have this capability - sorry.
