### 2025-09-09 #8853 @genomicssport
@JMencius Brilliant and sure, I will go through it and check and will incorporate which can make a improvement. @jromanowska Reference check summary (note 'MISSING' DOIs are suggestions that need verification):

✅ OK DOIs

- 10.1186/s13059-016-0974-4 is OK
- 10.1093/nar/gkq603 is OK
- 10.18129/B9.bioc.VariantAnnotation is OK
- 10.1093/nar/gkac931 is OK

🟡 SKIP DOIs

- None

❌ MISSING DOIs

- None

❌ INVALID DOIs

- None
