### 2025-09-09 #8853 @jromanowska
@genomicssport, I read the paper once more and I need you to revise it thoroughly. I've made some comments in the pull request. Your tool is very useful, but it's difficult to understand what it can do from the paper. Please, use some time on rewriting the paper, perhaps ask another person from the group to read it - it's always easier to spot places that need improvement when it's only you working on a task.
