### 2025-09-09 #8853 @genomicssport
@jromanowska Thats very good and i revised all those sentences as if you cant understand means no one can understand, so it needed revision. Well said and i read the paper and then i revised those sentence just now and added the terms and the explanation so that it becomes easier for everyone to understand.

I believe if it doesnt apply on me, it will not be applicable to anyone. So i took that and revised those sentences. Thank you so much. I must acknowledge you.
