### 2025-09-09 #8853 @genomicssport
@jromanowska I added you as [acknowledgement] and you can generate the current version and read that along. It addressed everything. You can download the latest build from here: https://github.com/genomicssport/eVaiutilities/actions/runs/17581461312
