### 2025-09-09 #8853 @genomicssport
@jromanowska thank you and here is the last version with the correction of the typo. https://github.com/genomicssport/eVaiutilities/actions/runs/17582234928.

Thank you all and i am very glad to have such interaction.
