### 2025-09-10 #8853 @jromanowska
@genomicssport, I appreciate your efforts, but still, the text needs a re-write, and you did not format the URLs as I've suggested (to be clickable). Please, have someone else in the co-author team read it and comment. _Use your time to make the text good._
