### 2025-09-10 #8853 @genomicssport
@jromanowska thank you and i will make the link clickable in a short while and i will read through the same. I think the paper reads very well now. I dont see a point. If you have any then can you specify me to the specific area that would like me to read. I would appreciate being specific rather than a general comment. How you think about that as i believe that it is a better approach to reach the solution. Dont you think so.
