### 2025-09-10 #8853 @genomicssport
@jromanowska I have finished making the links clickable and kindly let me know what to improve. The latest version can be found from here: https://github.com/genomicssport/eVaiutilities/actions/runs/17612091732. All links are clickable and i thank you for indicating this.

For next submission, i will take care of these ahead of the submission of the publication and hence all these comments which were meaningful well appreciated. If everything is OK, i would be pleased to have the acceptance of the paper.
