### 2025-09-10 #8853 @erik-whiting
> While I have not provided further comments at this stage, this should not be interpreted as indicating that the manuscript is free of issues. I have corrected obvious typographical and expression errors, but further improvements will need to be made by the authors themselves. And [@erik-whiting](https://github.com/erik-whiting) please don't AFK any more.

What do you mean "don't AFK any more"? My review is completed and I was stepping back to let the editor do their part.

> @erik-whiting You must be an English native speaker. If you are fine with this article, then I have no question.

I checked the box for **Quality of writing** because I was able to understand the article, despite the language barrier. I was taught not to comment on things like grammar and spelling, as this can come off insensitive, and to leave such things to the editor.

Having said that, the article is indeed difficult to read. I agree with the editor's suggestion that the author should run the text of the paper through ChatGPT or have a coworker look at it. The author is quite resistant to this suggestion though and is becoming inexplicably combative, so I have not participated in the discussion, since it seems like the editor is handling it well.
