### 2025-09-10 #8853 @JMencius
> I suggest GPT thank you.

That is exactly what I mean "don't AFK any more".
