### 2025-09-10 #8853 @jromanowska
@JMencius, thanks for the comment. I think your comment about @erik-whiting was a bit too harsh, but I assume you just wanted a second opinion from the other reviewer. Thank you for backing me up on this, @erik-whiting.

I will then wait, @genomicssport, until your co-authors have read and commented on the paper. Please, let us know.
