### 2025-09-10 #8853 @genomicssport
@jromanowska thank you for being correct in your approach and @JMencius thank you for apologizing. Me or any one has no right to mix academic or anything with personal relations. So please be polite on what you say and take correct decision. As an academic person or a researcher or a professional or professor your role is to see the respective field in that range and there is no space for bringing person words or relations. @jromanowska thank you the comment abd everyone has seen and read the paper. I am the lead author and developer for this. Everyone took part in revising, polishing and also suggesting new features. Please kindly let me know if you need any more information or I will wait for the acceptance.
