### 2025-09-13 #8853 @Kevin-Mattheus-Moerman
@genomicssport As you can see I've paused this submission at this point. I have just parsed the entire thread above and see several very disrespectful comments from your side e.g. https://github.com/openjournals/joss-reviews/issues/8853#issuecomment-3280576558

> I want a succinct response and not dabbling words.

 and

>  I hope you have written scientific paper or even reviewed before.

and

> she is not understanding. Can you help me with another editor as the review has already been done

and

> Please dont make academic and scientific community a place to collect rubbish language and personal communications.

There are other comments that are not professional in my opinion but the above summary is sufficient. I want to strongly stress that the handling editor @jromanowska is a very experienced editor and all our editors that have now review your paper and these comments have agreed that she has been doing a fantastic job editing here despite disrespectful comments from your side. I have paused this submission at this point, and expect an apology to the editor and reviewers from you on the above before we can proceed. If you disagree with my assessment then I suggest we reject this submission at this point.

Secondly, the reviewers and several editors that have now looked at your paper, all agreed the paper needed work and improvements. As such you should work really hard to understand the editor's/reviewer's points rather than be disrespectful. I see no reason here to assign a different editor. This is not an isolated editor's opinion we are talking about, all readers here agree. You simply need to work to improve the paper and implement exactly the changes requested. Failure to do so will result in the rejection of this paper.

Thirdly,  myself and @rkurchin are the track editor here. Please do not involve the other editors like Dan or Arfon. Address all communication to your handling editor @jromanowska. If you need to address the track AEiC's you can tag me or Rachel.

I would like to point out that this review is not completed until all steps are completed including the points raised in the post-review checklist and any additional changes requested by @jromanowska. As such please refrain from asking to proceed to acceptance. This is not something the authors can request. You have to satisfy all issues raised and then we will determine if we can proceed to acceptance.

Finally, the handling editor has asked you a simple question on several occasions which has, as far as I can see, not been answered yet, please do so. She asked:

> Have all the co-authors had time to read it and accept?

This is important to verify so please respond.
