### 2025-09-13 #8853 @genomicssport
@Kevin-Mattheus-Moerman Thank you for the response and i am the sole author of the paper. @jromanowska I am waiting patiently as i dont have to put pressure.

@jromanowska Sure, here it goes, I apologize if i have said something wrong, which might not be the case, but asking an apology is not a big thing. It might be possible that someone has read and misunderstood the language and hence asked for an apology.

So here you go, please accept my apologies for if anything has disrespected you. I also got disrespect from the editor or the reviewers but i dont need that as i am quite good in my thinking.

@Kevin-Mattheus-Moerman anything else that you would me to answer.
