### 2025-09-13 #8853 @erik-whiting
> @erik-whiting thanks for your review efforts here!
>
> I see you said your review was complete. However, several of your boxes are unticked. Can you or @jromanowska clarify if these are left blank on purpose or can now be ticked? Thanks

Apologies for the oversight, I have updated my checklist.
